feat: export MessageUnpackStream class
Export MessageUnpackStream class from the index.js, so once the package is required, you can create instances of both IOSDeviceLib and MessageUnpackStream. Move the implementation of IOSDeviceLib to a separate file to simplify the entry point.
