+++ typings/ios-device-lib.d.ts
@@ -2,4 +2,5 @@
 
 declare module "ios-device-lib" {
 	export const IOSDeviceLib: IOSDeviceLib.IOSDeviceLib;
+	export const MessageUnpackStream: IOSDeviceLib.MessageUnpackStream;
 }
